+++ R/WAASB.R
@@ -115,7 +115,7 @@
 #'   matrix of BLUPs for genotype-environment interaction, thus allowing the
 #'   computation of the WAASB index.
 #' @references
-#' Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, V.S. Marchioro, V.Q. de
+#' Olivoto, T., A.D.C. Lúcio, J.A.G. da silva, V.S. Marchioro, V.Q. de
 #' Souza, and E. Jost. 2019. Mean performance and stability in multi-environment
 #' trials I: Combining features of AMMI and BLUP techniques. Agron. J.
 #' 111:2949-2960.