+++ R/cv_blup.R
@@ -79,7 +79,7 @@
 #' **RMSPD**: A vector with nboot-estimates of the root mean squared
 #' prediction difference between predicted and validating data. *
 #' **RMSPDmean** The mean of RMSPDmean estimates.
-#' @references Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, V.S. Marchioro,
+#' @references Olivoto, T., A.D.C. Lúcio, J.A.G. da silva, V.S. Marchioro,
 #'   V.Q. de Souza, and E. Jost. 2019. Mean performance and stability in
 #'   multi-environment trials I: Combining features of AMMI and BLUP techniques.
 #'   Agron. J. 111:2949-2960.