+++ R/ge_stats.R
@@ -112,7 +112,7 @@
 #'   performance for cultivar x location data. Can. J. Plant Sci. 68:193-198.
 #'   \doi{10.4141/cjps88-018}
 #'
-#'   Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, V.S. Marchioro, V.Q. de
+#'   Olivoto, T., A.D.C. Lúcio, J.A.G. da silva, V.S. Marchioro, V.Q. de
 #'   Souza, and E. Jost. 2019a. Mean performance and stability in
 #'   multi-environment trials I: Combining features of AMMI and BLUP techniques.
 #'   Agron. J. 111:2949-2960. \doi{10.2134/agronj2019.03.0220}
@@ -145,7 +145,7 @@
 #'
 #' @export
 #' @seealso [acv()], [ammi_indexes()], [ecovalence()], [Fox()], [gai()],
-#'   [ge_reg()], [hmgv()], [hmrpgv()], [rpgv()], [Huehn()], [polar()],
+#'   [ge_reg()], [hmgv()], [hmrpgv()], [rpgv()], [Huehn()], [ge_polar()],
 #'   [Shukla()], [superiority()], [Thennarasu()], [waas()], [waasb()]
 #'
 #' @examples