+++ R/get_model_data.R
@@ -82,6 +82,13 @@
 #' * `"crossloads"` The canonical cross-loadings.
 #' * `"canonical"` The canonical correlations and hypothesis testing.
 #'
+#'  **Objects of class `colindiag`:**
+#' * `"cormat"` The correlation matrix betwen predictors.
+#' * `"corlist"` The correlations in a 'long' format
+#' * `"evalevet"` The eigenvalue with associated eigenvectors
+#' * `"VIF"` The Variance Inflation Factor
+#' * `"indicators"` The colinearity indicators
+#'
 #'  **Objects of class `ecovalence`:**
 #' * `"Ecoval"` Ecovalence value (default).
 #' * `"Ecov_perc"` Ecovalence in percentage value.
@@ -279,13 +286,13 @@
 #' Huehn, V.M. 1979. Beitrage zur erfassung der phanotypischen stabilitat. EDV
 #' Med. Biol. 10:112.
 #'
-#' Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, V.S. Marchioro, V.Q. de
+#' Olivoto, T., A.D.C. Lúcio, J.A.G. da silva, V.S. Marchioro, V.Q. de
 #' Souza, and E. Jost. 2019a. Mean performance and stability in
 #' multi-environment trials I: Combining features of AMMI and BLUP techniques.
 #' Agron. J. 111:2949-2960.
 #' \doi{10.2134/agronj2019.03.0220}
 #'
-#' Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, B.G. Sari, and M.I. Diel.
+#' Olivoto, T., A.D.C. Lúcio, J.A.G. da silva, B.G. Sari, and M.I. Diel.
 #' 2019b. Mean performance and stability in multi-environment trials II:
 #' Selection based on multiple traits. Agron. J. 111:2961-2969.
 #' \doi{10.2134/agronj2019.03.0221}
@@ -369,7 +376,7 @@ get_model_data <- function(x,
                       "gafem", "gafem_group", "gamem_group", "anova_ind", "gge", "can_cor",
                       "can_cor_group", "gytb", "ge_acv", "ge_polar", "mgidi", "mtsi",
                       "env_stratification", "fai_blup", "sh", "mps", "mtmps", "path_coeff",
-                      "path_coeff_seq", "group_path_seq", "group_path"))) {
+                      "path_coeff_seq", "group_path_seq", "group_path", "colindiag", "colingroup"))) {
     stop("Invalid class in object ", call_f[["x"]], ". See ?get_model_data for more information.", call. = FALSE)
   }
   if (!is.null(what) && what != "PCA" && substr(what, 1, 2) == "PC") {
@@ -436,6 +443,28 @@ get_model_data <- function(x,
                "sense_mper", "stab_method", "wstab", "sense_stab")
   check33 <- c("coef", "eigenval", "vif")
   check34 <- c("resp_fc", "resp_sc", "resp_sc2", "fc_sc_coef")
+  check35 <- c("cormat", "corlist", "evalevet", "VIF", "indicators")
+
+
+  if(has_class(x, c("colindiag", "colingroup"))){
+    if (is.null(what)){
+      what <- "indicators"
+    }
+    if (!what %in% check35) {
+      stop("Invalid value in 'what' for object of class 'colindiag'. Allowed are ", paste(check35, collapse = ", "), call. = FALSE)
+    }
+    if(has_class(x, "colingroup")){
+      bind <-
+        x %>%
+        mutate(data = map(data, ~.x %>% .[[what]])) %>%
+        unnest(data)
+    } else{
+      bind <- x[[what]]
+    }
+  }
+
+
+
   if(has_class(x, c("path_coeff", "group_path"))){
     if (is.null(what)){
       what <- "coef"