+++ R/mps.R
@@ -103,7 +103,7 @@
 #' methods for selecting stable and adapted durum wheat genotypes in variable
 #' environments. Euphytica, 159(3), 419-432. \doi{10.1007/s10681-007-9600-6}
 #'
-#' Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, V.S. Marchioro, V.Q. de
+#' Olivoto, T., A.D.C. Lúcio, J.A.G. da silva, V.S. Marchioro, V.Q. de
 #' Souza, and E. Jost. 2019. Mean performance and stability in multi-environment
 #' trials I: Combining features of AMMI and BLUP techniques. Agron. J.
 #' \doi{10.2134/agronj2019.03.0220}
@@ -585,7 +585,8 @@ mps <- function(.data,
       sense_mper = ifelse(sense_mper == 0, "l", "h"),
       stab_method = stability,
       wstab = peso_stab,
-      sense_stab = ifelse(rescaled == 0, "l", "h")
+      sense_stab = ifelse(rescaled == 0, "l", "h"),
+      .resid =  gmd(mod, "data", verbose = FALSE) %>% mean_by(ENV, GEN)
     ) %>% set_class("mps")
   )
 }