+++ R/mtmps.R
@@ -64,7 +64,7 @@
 #' @md
 #' @author Tiago Olivoto \email{tiagoolivoto@@gmail.com}
 #' @export
-#' @references Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, B.G. Sari, and
+#' @references Olivoto, T., A.D.C. Lúcio, J.A.G. da silva, B.G. Sari, and
 #'   M.I. Diel. 2019. Mean performance and stability in multi-environment trials
 #'   II: Selection based on multiple traits. Agron. J. 111:2961-2969.
 #' \doi{10.2134/agronj2019.03.0220}
@@ -378,7 +378,7 @@ mtmps <- function(model,
 #' @author Tiago Olivoto \email{tiagoolivoto@@gmail.com}
 #' @method plot mtmps
 #' @export
-#' @references Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019. Mean performance and stability in multi-environment trials II: Selection based on multiple traits. Agron. J. (in press).
+#' @references Olivoto, T., A.D.C. Lúcio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019. Mean performance and stability in multi-environment trials II: Selection based on multiple traits. Agron. J. (in press).
 #' @examples
 #' \donttest{
 #' library(metan)