+++ R/utilities.R
@@ -1067,6 +1067,7 @@ cv <- function(.data, ..., na.rm = FALSE) {
 #' @name utils_stats
 #' @export
 freq_table <- function(.data, var, k = NULL, digits = 3){
+  var <- rlang::ensym(var)
   if(is_grouped_df(.data)){
     res <-
       metan::doo(.data,
@@ -1080,8 +1081,6 @@ freq_table <- function(.data, var, k = NULL, digits = 3){
       res %>%
       mutate(freqs = map(data, ~.x %>% .[["breaks"]])) %>%
       remove_cols(data)
-    list_breaks <- breaks$freqs
-    names(list_breaks) <- breaks$cor_grao
     return(list(freqs = freqs,
                 breaks = breaks))
 
@@ -1486,10 +1485,16 @@ row_col_mean <- function(.data, na.rm = FALSE) {
     stop("All columns in '.data' must be numeric")
   }
   mat <- as.matrix(.data)
-  row_means <- rowMeans(mat, na.rm = na.rm)
-  col_means <- colMeans(mat, na.rm = na.rm)
-  cmeans <- suppressWarnings(cbind(mat,  row_means) %>% rbind(col_means))
-  rownames(cmeans) <- c(1:nrow(mat), "col_means")
+
+  if(is.null(rownames(mat))){
+    row_names <- 1:nrow(mat)
+  } else{
+    row_names <- rownames(mat)
+  }
+  row_mean <- rowMeans(mat, na.rm = na.rm)
+  col_mean <- colMeans(mat, na.rm = na.rm)
+  cmeans <- suppressWarnings(cbind(mat,  row_mean) %>% rbind(col_mean))
+  rownames(cmeans) <- c(row_names, "col_mean")
   cmeans[nrow(cmeans), ncol(cmeans)] <- mean(mat, na.rm = na.rm)
   return(cmeans)
 }
@@ -1505,10 +1510,15 @@ row_col_sum <- function(.data, na.rm = FALSE) {
     stop("All columns in '.data' must be numeric")
   }
   mat <- as.matrix(.data)
-  row_sums <- rowSums(mat, na.rm = na.rm)
-  col_sums <- colSums(mat, na.rm = na.rm)
-  cmeans <- suppressWarnings(cbind(mat,  row_sums) %>% rbind(col_sums))
-  rownames(cmeans) <- c(1:nrow(mat), "col_sums")
+  if(is.null(rownames(mat))){
+    row_names <- 1:nrow(mat)
+  } else{
+    row_names <- rownames(mat)
+  }
+  row_sum <- rowSums(mat, na.rm = na.rm)
+  col_sum <- colSums(mat, na.rm = na.rm)
+  cmeans <- suppressWarnings(cbind(mat,  row_sum) %>% rbind(col_sum))
+  rownames(cmeans) <- c(row_names, "col_sum")
   cmeans[nrow(cmeans), ncol(cmeans)] <- sum(mat, na.rm = na.rm)
   return(cmeans)
 }