+++ R/wsmp.R
@@ -37,7 +37,7 @@
 #' @param progbar A logical argument to define if a progress bar is shown.
 #'   Default is `TRUE`.
 #' @references
-#' Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, V.S. Marchioro, V.Q. de
+#' Olivoto, T., A.D.C. Lúcio, J.A.G. da silva, V.S. Marchioro, V.Q. de
 #' Souza, and E. Jost. 2019. Mean performance and stability in multi-environment
 #' trials I: Combining features of AMMI and BLUP techniques. Agron. J.
 #' \doi{10.2134/agronj2019.03.0220}
@@ -200,7 +200,7 @@ wsmp <- function(model,
             if (progbar == TRUE) {
               run_progress(pb,
                            actual = initial,
-                           text = paste("Ranks considering ", PesoResp, " for GY and ", PesoWAAS, " for WAASB", sep = ""))
+                           text = paste("Ranks considering ", PesoResp, " for Y and ", PesoWAAS, " for WAASB", sep = ""))
 
             }
           }